backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-secret-manager==2.16.4
google-cloud-firestore==2.13.1
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
aioredis==2.0.1
python-dotenv==1.0.0
corsheaders==4.0.0
passlib[bcrypt]==1.7.4